requirements.txt
gradio
scikit-learn
numpy
joblib
pillow
